tests: move `batch_add_schnorrsig` tests to schnorrsig module
