+++ src/tests.c
@@ -6870,6 +6870,9 @@ void run_ecdsa_edge_cases(void) {
 
 #ifdef ENABLE_MODULE_SCHNORRSIG
 # include "modules/schnorrsig/tests_impl.h"
+# ifdef ENABLE_MODULE_BATCH
+#  include "modules/schnorrsig/batch_add_tests_impl.h"
+# endif
 #endif
 
 #ifdef ENABLE_MODULE_BATCH
@@ -7174,6 +7177,9 @@ int main(int argc, char **argv) {
 
 #ifdef ENABLE_MODULE_SCHNORRSIG
     run_schnorrsig_tests();
+# ifdef ENABLE_MODULE_BATCH
+    run_batch_add_schnorrsig_tests();
+# endif
 #endif
 
 #ifdef ENABLE_MODULE_BATCH
